2부/22장/flexible/requirements.txt
Flask==0.12.1
gunicorn==19.7.1
google-cloud==0.24.0
